Do datetimes properly using a datetime library
Previously I had been manually formatting the js.Date objects, but this is not conducive to being able to write code that handles timezones and daylight saving time. cljs-time seems to be the standard and works off the goog.date.DateTime objects which are already part of the compilation flow.